api rework

+++ WebApi/Controllers/UsersController.cs
@@ -4,12 +4,14 @@ using System.Diagnostics;
 using System.Linq;
 using System.Security.Cryptography;
 using System.Threading.Tasks;
+using AutoMapper;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using WebApi.Handlers;
 using WebApi.Models;
+using WebApi.Models.DataTransferObjects;
 
 
 namespace WebApi.Controllers
@@ -20,66 +22,81 @@ namespace WebApi.Controllers
     {
         private readonly WebApiContext _context;
         private readonly ISalter _salter;
+        private readonly IMapper _mapper;
         private SaltAgorithm _saltAgorithm;
 
-        public UsersController(WebApiContext context, ISalter salter)
+        public UsersController(WebApiContext context, ISalter salter, IMapper mapper)
         {
             _context = context;
             _salter = salter;
-        }
-
-        // GET: api/Users
-        [HttpGet]
-        public async Task<ActionResult<IEnumerable<User>>> GetUser()
-        {
-            return await _context.Users.ToListAsync();
-        }
-
-        // GET: api/Users/5
-        [HttpGet("{id}")]
-        public async Task<ActionResult<User>> GetUser(int id)
-        {
-            var user = await _context.Users.FindAsync(id);
-
-            if (user == null)
-            {
-                return NotFound();
-            }
-
-            return user;
+            _mapper = mapper;
         }
         
         // GET: api/Users/Authorize
         [HttpGet("Authorize")]
         [Authorize]
-        public async Task<ActionResult<bool>> CheckAuthorization()
+        public bool CheckAuthorization()
         {
             return true;
         }
-        // POST: api/Users/RegisterUser
-        [HttpPost("RegisterUser")]
-        public async Task<ActionResult<User>> RegisterUser([FromBody] User user, string password)
+
+        // GET: api/Users
+        [HttpGet]
+        public async Task<ActionResult<List<UserDTO>>> GetUsers()
+        {
+            return _mapper.Map<List<UserDTO>>(await _context.Users.ToListAsync());
+        }
+
+        // GET: api/Users/5
+        [HttpGet("{id}")]
+        public async Task<ActionResult<UserDTO>> GetUser(int id)
+        {
+            var user = await _context.Users.FindAsync(id);
+
+            if (user == null)
+            {
+                return NotFound("User not found");
+            }
+
+            return _mapper.Map<UserDTO>(user);
+        }
+
+        // POST: api/Users
+        [HttpPost]
+        public async Task<ActionResult<UserDTO>> RegisterUser([FromBody] UserDTO userDto, string password)
         {
             try
             {
-                if (string.IsNullOrEmpty(user.Name))
-                    throw new ArgumentNullException(nameof(user.Name));
+                if (string.IsNullOrEmpty(userDto.Name))
+                {
+                    throw new ArgumentNullException(nameof(userDto.Name));
+                }
+                
                 if (string.IsNullOrEmpty(password))
+                {
                     throw new ArgumentNullException(nameof(password), "cannot be null or empty");
-                if (CheckIfUsernameTaken(user.Name))
+                }
+                if (await CheckIfUsernameTaken(userDto.Name))
+                {
                     return StatusCode(406);
+                }
+                
                 var salt = _salter.CreateSalt();
-                if(_saltAgorithm == null)
-                    _saltAgorithm = new SaltAgorithm();
+                _saltAgorithm ??= new SaltAgorithm();
                 BasicAuthenticationHandler.Hashing hashdelegate = _saltAgorithm.Hash;
                 var hash = _salter.GenerateSaltedHash(password, salt,hashdelegate);
-                _context.Users.Add(user);
-                _context.SaveChanges();
-                var fetcheduser = _context.Users.FirstOrDefault(u => u.Name == user.Name);
-                var userPassword = new Password {Hash = hash, UserId = fetcheduser.Id, Salt = salt};
+
+                var user = _mapper.Map<User>(userDto);
+                
+                await _context.Users.AddAsync(_mapper.Map<User>(user));
+                await _context.SaveChangesAsync();
+                
+                var userPassword = new Password {Hash = hash, UserId = user.Id, Salt = salt};
+                
                 await _context.Passwords.AddAsync(userPassword);
                 await _context.SaveChangesAsync();
-                return CreatedAtAction("GetUser", new {id = user.Id}, user);
+                
+                return CreatedAtAction("GetUser", new {id = user.Id}, _mapper.Map<UserDTO>(user));
             }
             catch (ArgumentNullException e)
             {
@@ -87,9 +104,9 @@ namespace WebApi.Controllers
             }
             
         }
-        private bool CheckIfUsernameTaken(string username)
+        private async Task<bool> CheckIfUsernameTaken(string username)
         {
-            return _context.Users.Count(user => user.Name == username) > 0;
+            return (await _context.Users.CountAsync(user => user.Name == username)) > 0;
         }
 
 
@@ -97,13 +114,20 @@ namespace WebApi.Controllers
         // To protect from overposting attacks, enable the specific properties you want to bind to, for
         // more details, see https://go.microsoft.com/fwlink/?linkid=2123754.
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutUser(int id, User user)
+        [Authorize]
+        public async Task<IActionResult> PutUser(int id, UserDTO userDto)
         {
-            if (id != user.Id)
+            if (id != userDto.Id)
             {
                 return BadRequest();
             }
 
+            if (id != AuthenticatedUserId())
+            {
+                return Unauthorized("You are not this user");
+            }
+
+            var user = _mapper.Map<User>(userDto);
             _context.Entry(user).State = EntityState.Modified;
 
             try
@@ -116,46 +140,42 @@ namespace WebApi.Controllers
                 {
                     return NotFound();
                 }
-                else
-                {
-                    throw;
-                }
+                throw;
             }
-
             return NoContent();
-        }
-
-        // POST: api/Users
-        // To protect from overposting attacks, enable the specific properties you want to bind to, for
-        // more details, see https://go.microsoft.com/fwlink/?linkid=2123754.
-        [HttpPost]
-        public async Task<ActionResult<User>> PostUser(User user)
-        {
-            _context.Users.Add(user);
-            await _context.SaveChangesAsync();
-
-            return CreatedAtAction("GetUser", new { id = user.Id }, user);
         }
 
         // DELETE: api/Users/5
         [HttpDelete("{id}")]
-        public async Task<ActionResult<User>> DeleteUser(int id)
+        [Authorize]
+        public async Task<ActionResult<UserDTO>> DeleteUser(int id)
         {
             var user = await _context.Users.FindAsync(id);
+            
             if (user == null)
             {
-                return NotFound();
+                return NotFound("User not found");
+            }
+
+            if (id != AuthenticatedUserId())
+            {
+                return Unauthorized("You are not this user");
             }
 
             _context.Users.Remove(user);
             await _context.SaveChangesAsync();
 
-            return user;
+            return _mapper.Map<UserDTO>(user);
         }
 
         private bool UserExists(int id)
         {
             return _context.Users.Any(e => e.Id == id);
+        }
+        
+        private int AuthenticatedUserId()
+        {
+            return int.Parse(HttpContext.User.Identity.Name ?? "-1");
         }
         
     }
